WithTemporalFootprint.with_temporal_borders_replaced
The method is now part of the WithTemporalFootprint interface.
It was not the case before due to circular imports due to typing and the
fact that I did not understand that if TYPE_CHECKING could be used for
that purpose.
