+++ src/tqec/compile/blocks/layers/temporal.py
@@ -1,8 +1,15 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING, Mapping
 
+from tqec.compile.blocks.enums import TemporalBlockBorder
 from tqec.utils.scale import LinearFunction
+
+if TYPE_CHECKING:
+    # Avoid circular dependency just because of typing annotations.
+    from tqec.compile.blocks.layers.atomic.base import BaseLayer
+    from tqec.compile.blocks.layers.composed.base import BaseComposedLayer
 
 
 class WithTemporalFootprint(ABC):
@@ -35,3 +42,21 @@ class WithTemporalFootprint(ABC):
             the provided scaling parameter ``k``.
         """
         return self.scalable_timesteps.integer_eval(k)
+
+    @abstractmethod
+    def with_temporal_borders_replaced(
+        self, border_replacements: Mapping[TemporalBlockBorder, BaseLayer | None]
+    ) -> BaseLayer | BaseComposedLayer | None:
+        """Returns ``self`` with the provided temporal borders replaced.
+
+        Args:
+            borders: a mapping from temporal borders to replace to their
+                replacement. A value of ``None`` as a replacement means that the
+                border is removed.
+
+        Returns:
+            a copy of ``self`` with the provided ``borders`` replaced, or ``None``
+            if replacing the provided ``borders`` from ``self`` result in an
+            empty temporal footprint.
+        """
+        pass
